Dot Voting/VotingCard: Always clamp to 3 lines
The idea was to make more space for the "N more" badge that shows up
when there are more than 2 options to vote for. Maybe that made sense
when this logic was added, but the "N more" badge no longer adjusts the
vertical position of the options, so the number of description lines
need no adjustment.